menuchecker 4.0 - Bug fix
Fixed an off-by-one bug that was moving the last item of each menu to
the next one (Covel -> De Neve -> Feast) and cutting off the last item
of Feast.

Also note that the script scrapes the UCLA mobile dining page and that
because that page is inconsistent, this script may give inconsistent
results as well.